customer adding dishes to one order

+++ src/main/java/io/swagger/repository/DishRepository.java
@@ -1,7 +1,8 @@
 package io.swagger.repository;
 
 import io.swagger.entity.Dish;
+import io.swagger.models.auth.In;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DishRepository extends JpaRepository<Dish,Long> {
+public interface DishRepository extends JpaRepository<Dish, Integer> {
 }
